hospital-management-system/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
pandas
numpy
joblib
python-dotenv
supabase
httpx
torch
torchvision
Pillow
scikit-image
xgboost
prophet
statsmodels
